Refactor util class into dedicated file
